removed code duplication
